Massive refactor and Graphics Library Added
- Refactored a lot of code
- Imported, modified, and fixed Mode13h code from previous project
  (VGA initialized using BIOS interrupts instead of VGA registers and
  code has been attempted for Real Mode)
- gl/ now holds a graphics library, providing basic user-level graphical
  functions

+++ src/usr/hellow.c
@@ -7,7 +7,7 @@
 */
 
 /** Headers    **/
-#include "gcc16.h"
+#include "../kern/gcc16.h"
 #include "hellow.h"
 
 /*
@@ -26,7 +26,7 @@ void hellow_init(Program* prog)
 /*
 ** Main method for hello world program
 */
-uint16_t hellow_main(uint16_t argc, char* argv[])
+uint8_t hellow_main(uint8_t argc, char* argv[])
 {
     kio_print("Hello, world!\n");
     return EXIT_SUCCESS;
